Added texture configuration, loading and texture mapping to walls
main.cpp: Cleaned the
code a bit, added texture configuration call

common.hpp: New files that contains all the imports and flag definitios

grid.hpp: Fixed ambiguous call to sqrt

room.hpp: Removed unecessary import

wall.hpp: Added code to render with textures instead of vertex colors

textures.hpp: New file to deal with texture configuration and loading from a given file path

textures/: New directory that contains all textures used

models/toyplane.obj: New 3D model added

+++ grid.hpp
@@ -6,7 +6,6 @@
 #endif
 
 #define GRID_DEBUG 0
-#include <math.h>
 
 typedef struct gridNode 
 {
@@ -18,7 +17,7 @@ void addRoom (GridNode*, Room);
 
 GridNode (createGrid) (int elements, float roomSize)
 {
-	int dimension = ceil(sqrt((float)elements));
+	int dimension = ceil(sqrt(elements));
 	int elementsCreated = 0;
 	vec3 gridStart = vec3 (0.0f, 0.0f, 0.0f);
